Tests and refactoring
