+++ src/SharedGenerators/Extractor.cs
@@ -7,63 +7,62 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-namespace SpeedyGenerators
+namespace SpeedyGenerators;
+
+internal static class Extractor
 {
-    internal static class Extractor
+    public static string[] ExtractComments(FieldDeclarationSyntax fieldDeclaration)
     {
-        public static string[] ExtractComments(FieldDeclarationSyntax fieldDeclaration)
-        {
-            return ExtractComments(fieldDeclaration.GetLeadingTrivia());
-        }
+        return ExtractComments(fieldDeclaration.GetLeadingTrivia());
+    }
 
-        public static string[] ExtractComments(SyntaxTriviaList syntaxTrivias)
+    public static string[] ExtractComments(SyntaxTriviaList syntaxTrivias)
+    {
+        List<string> lines = new();
+        foreach (var trivia in syntaxTrivias)
         {
-            List<string> lines = new();
-            foreach (var trivia in syntaxTrivias)
+            if (trivia.IsKind(SyntaxKind.SingleLineDocumentationCommentTrivia))
             {
-                if (trivia.IsKind(SyntaxKind.SingleLineDocumentationCommentTrivia))
-                {
-                    var structure = trivia.GetStructure()
-                        as DocumentationCommentTriviaSyntax;
-                    if (structure == null) continue;
+                var structure = trivia.GetStructure()
+                    as DocumentationCommentTriviaSyntax;
+                if (structure == null) continue;
 
-                    foreach (var contentLine in structure.Content
-                        .OfType<XmlElementSyntax>())
+                foreach (var contentLine in structure.Content
+                    .OfType<XmlElementSyntax>())
+                {
+                    foreach (var xml in contentLine.Content)
                     {
-                        foreach (var xml in contentLine.Content)
-                        {
-                            lines.Add(xml.ToString()
-                                .Trim(new[] { '\r', '\n', '/', ' ' }));
-                        }
+                        lines.Add(xml.ToString()
+                            .Trim(new[] { '\r', '\n', '/', ' ' }));
                     }
                 }
             }
-
-            return lines.ToArray();
         }
 
-        public static MakePropertyArguments? ExtractMakePropertyArguments(
-            AttributeSyntax attribute)
-        {
-            var args = attribute.ArgumentList?.Arguments
-                .Select(a => a.ToString())
-                .ToArray();
+        return lines.ToArray();
+    }
 
-            MakePropertyArguments.TryParse(args, out MakePropertyArguments? arguments);
+    public static MakePropertyArguments? ExtractMakePropertyArguments(
+        AttributeSyntax attribute)
+    {
+        var args = attribute.ArgumentList?.Arguments
+            .Select(a => a.ToString())
+            .ToArray();
 
-            return arguments;
-        }
+        MakePropertyArguments.TryParse(args, out MakePropertyArguments? arguments);
 
-        public static MakeConcreteArguments? ExtractMakeConcreteArguments(
-            AttributeSyntax attribute)
-        {
-            var args = attribute.ArgumentList?.Arguments
-                .Select(a => a.ToString())
-                .ToArray();
+        return arguments;
+    }
 
-            MakeConcreteArguments.TryParse(args, out MakeConcreteArguments? arguments);
+    public static MakeConcreteArguments? ExtractMakeConcreteArguments(
+        AttributeSyntax attribute)
+    {
+        var args = attribute.ArgumentList?.Arguments
+            .Select(a => a.ToString())
+            .ToArray();
 
-            return arguments;
-        }
+        MakeConcreteArguments.TryParse(args, out MakeConcreteArguments? arguments);
+
+        return arguments;
     }
 }
